+++ internal/pluck/extract.go
