+++ internal/pluck/extract_test.go
